+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Fix Wine crash when the Steam application and Protontricks are running at the same time
  - Fix Steam installation detection when both non-Flatpak and Flatpak versions of Steam are installed for the same user
  - Fix Protontricks crash when Proton installation is incomplete
+ - Fix Protontricks crash when both Flatpak and non-Flatpak versions of Steam are installed
 
 ## [1.7.0] - 2022-01-08
 ### Changed